scripts: add timestamp to bust the cache to retrieve latest updates
On GitHub Pages the files are often so good cached that old / stale data
is still retrieved and not the latest version of `index.json`.

Even a reload with the "disable cache" option enabled in the network
settings of the browser developer tools does not help in most cases.

Thus we have to bust the cache for every load by adding a timestamp in
this case.

There are also other solutions but these would affect all AJAX calls and
this solution busts the cache for only the `index.json` file.