storage/github: add commit_message_suffix as config entry
Adding '[ci skip]' made sense in the original implementation, when CI
was not often used to act on a status repo. In the age of GitHub
Actions, it would be helpful to be able to control whether this string
is appended.

The current implementation leaves "[ci skip]" enabled for all users,
and the user can optionally change it by adding this setting to their
storage configuration:

	"commit_message_suffix": "user-defined-string"

If the user sets the suffix to a single space (" "), then it will be
disabled.